# contracts/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AuditManager PC-01 API",
    "version": "1.0.0-draft.1",
    "summary": "The exact PC-01 journey: one project, one AR PDF, one text-consistency run, one CSV.",
...
    "contact": {
      "name": "docs/program/P02_SEAMS.md - seam S8"
    },
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Version-prefixed base path. A breaking change is a new contract version, never an edit here."
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/execution/queue": {
      "get": {
        "tags": [
          "execution"
        ],
        "summary": "List Queue",
        "operationId": "listExecutionQueue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "description": "The durable queue and dispatch pause state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionQueuePage"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/execution/journal": {
      "get": {
        "tags": [
          "execution"
        ],
        "summary": "List Journal",
        "operationId": "listExecutionJournal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RunId"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "description": "The execution journal, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionJournalPage"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/runs/{run_id}/cancel": {
      "post": {
        "tags": [
          "execution"
        ],
        "summary": "Cancel Run",
        "operationId": "cancelRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "200": {
            "description": "The cancelled run or a replayed result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/StateConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/runs/{run_id}/reaudit": {
      "post": {
        "tags": [
          "execution"
        ],
        "summary": "Reaudit Run",
        "operationId": "reauditRun",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RunId"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "The newly accepted re-audit run or a replayed result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/StateConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/execution/queue/{job_id}/priority": {
      "put": {
        "tags": [
          "execution"
        ],
        "summary": "Set Priority",
        "operationId": "setJobPriority",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          },
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetJobPriorityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The queued job with its new priority.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionQueueItem"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/StateConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    },
    "/execution/dispatch": {
      "put": {
        "tags": [
          "execution"
        ],
        "summary": "Set Paused",
        "operationId": "setExecutionPaused",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetExecutionPausedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The durable dispatch pause state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDispatchStatus"
                }
              }
            },
            "headers": {
              "X-Correlation-Id": {
                "required": true,
                "schema": {
                  "$ref": "#/components/schemas/CorrelationId"
                },
                "description": "The correlation id for this request, assigned when the caller supplied none."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "409": {
            "$ref": "#/components/responses/StateConflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/DependencyUnavailable"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The authorization seam. Every operation in this document requires a bearer credential presented as `Authorization: Bearer <token>`, except `issueToken`, `submitRegistration` and `readRegistrationStatus`, whose own `security` is the empty requirement: the credential exchange because it is what a caller with no credential uses to obtain one, and the registration pair because an applicant has no account yet.\n\n**This declares the seam, not its implementation.** No issuer, no discovery URL, no flow and no token format appears here, and none may be added: the deployment decides them and the same operations must keep working when it does. What a subject may do is not carried by the scheme or the credential either: since `R-55` it is the account's role set (`Role`), which the server reads from the account on every request. `bearerFormat` is deliberately absent for that reason -- an opaque deployment token and an OIDC-issued JWT are both bearer credentials, and naming the format would pin the document to whichever one is current.\n\nA request with no credential, or one the deployment does not accept, is `401 authentication_required`; so is an exchange whose login and password the deployment does not accept, which is the same refusal and not a second one. A credential the deployment accepts, for a subject not permitted this operation on this resource, is `403 permission_denied`, its `required_capability` naming what is missing (`password_changed`, `profile_completed`, `role:expert` or `role:admin`); authorization is decided server-side and the response never reveals the existence of a resource the caller may not see, so absence and refusal are reported as this surface's `404 not_found` where revealing them would leak.\n\nNothing outside `/api/v1` is covered. The operational liveness and readiness plane answers on its own port, carries no product meaning and needs no credential."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Client-supplied command key. Passed through to the owning command handler unchanged and never re-derived in the router. Never returned in a response body and never placed in an error envelope: it is a forbidden detail key.",
        "schema": {
          "$ref": "#/components/schemas/IdempotencyKey"
        }
      },
      "CorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "description": "Correlates this request with its diagnostic record. The edge assigns one when the caller does not supply it.",
        "schema": {
          "$ref": "#/components/schemas/CorrelationId"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque continuation token from the previous page's `next_cursor`. Never parsed by a client and never constructed by one.",
        "schema": {
          "$ref": "#/components/schemas/Cursor"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Page size.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
...
      "RunId": {
        "name": "run_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/RunId"
        }
      },
...
      "JobId": {
        "name": "job_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/JobId"
        }
      }
    },
...
    "responses": {
      "AuthenticationRequired": {
        "description": "`authentication_required`. No valid authenticated subject was presented: the request carried no bearer credential, or one this deployment does not accept, or -- on the credential exchange, which presents no credential by design -- a login and password pair this deployment does not accept. The response carries no hint about the addressed resource, so it is the same answer whether or not the resource exists.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "PermissionDenied": {
        "description": "`permission_denied`. The authenticated subject is not permitted to perform this operation on this resource. Authorization is decided server-side. This is the caller's rights and nothing else: a dependency that refuses the *application's* own credential is `dependency_credential_refused` on the 500 response, which is a server fault and not this one.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "The addressed aggregate does not exist or is not visible. The response never reveals the existence of a resource the caller may not see.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "The request or payload violates a declared schema, enum, format or invariant. Nothing was created or changed.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An unclassified server fault, a declared run failure, or `dependency_credential_refused` -- a dependency refusing the application's own credential, which is a server-side configuration fault with no authenticated subject in it. The original internal code stays in protected diagnostics and never reaches this envelope.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "DependencyUnavailable": {
        "description": "`dependency_unavailable`, or since `R-8` `staged_upload_lost` on a write that stages bytes. The caller retries; nothing was partially applied. `idempotency_key_in_progress` is also retryable and is reachable on every write, so a caller reads `retryable` from the envelope, where the catalog pins it per code, and never infers it from the status or from a description like this one.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "StateConflict": {
        "description": "The change conflicts with durable state. `conflict` carries `details.conflict_reason`: `login_taken` (an active account holds the login), `request_pending` (a request for the login is already pending), `queue_full` (the pending queue is at its cap of 100), `account_referenced` (the account is referenced and cannot be purged) or `last_admin` (no active account would hold `admin`). `state_transition_not_allowed` names the `app_user` or `registration_request` machine with `details.machine`, `current_state` and `requested_state`. On an operation that takes an idempotency key the idempotency-key codes are reported here too.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "RunId": {
        "type": "string",
        "pattern": "^run_[0-9A-HJKMNP-TV-Z]{26}$",
        "examples": [
          "run_01M2545JSD15ETSNNV904X991K"
        ]
      },
...
      "AnalysisProfileId": {
        "type": "string",
        "pattern": "^ap_[0-9A-HJKMNP-TV-Z]{26}$"
      },
...
      "CorrelationId": {
        "type": "string",
        "description": "Not an entity identity. Never a foreign key and authorizes nothing.",
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
        "minLength": 1,
        "maxLength": 128
      },
...
      "Cursor": {
        "type": "string",
        "description": "Opaque pagination token. It encodes the sort key of the last item on the page and nothing else; it is never a database row number or sequence value.",
        "minLength": 1,
        "maxLength": 512
      },
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "next_cursor"
        ],
        "properties": {
          "next_cursor": {
            "description": "Token for the next page, or null when this is the last page.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Cursor"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "RunStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "run_id",
          "project_uid",
          "version_uid",
          "state",
          "provider_mode",
          "stages",
          "created_at"
        ],
        "properties": {
          "run_id": {
            "$ref": "#/components/schemas/RunId"
          },
          "project_uid": {
            "$ref": "#/components/schemas/ProjectUid"
          },
          "version_uid": {
            "$ref": "#/components/schemas/VersionUid"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "provider_mode": {
            "$ref": "#/components/schemas/ProviderMode"
          },
          "analysis_profile_id": {
            "$ref": "#/components/schemas/AnalysisProfileId"
          },
          "prompt_bundle_id": {
            "$ref": "#/components/schemas/PromptBundleId"
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageState"
            }
          },
          "degradation_set": {
            "type": "array",
            "description": "The explicitly recorded missing or degraded stage set. Non-empty exactly when the run is `partial`; a `published` run carries none.",
            "items": {
              "$ref": "#/components/schemas/StageId"
            }
          },
          "terminal_reason": {
            "description": "The catalog code a `failed` run terminated with. Null otherwise.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "terminal_detail": {
            "type": "object",
            "description": "D-46: safe scalar classifiers saying WHICH dependency a `failed` run terminated on, restricted to the `safe_detail_keys` the domain catalog declares for the code in `terminal_reason`. Present only beside a `terminal_reason`, because the allowlist that bounds these keys is a property of the reported code; absent when there is nothing to say, and never an empty object. The same shape and the same rule as `ErrorEnvelope.details`, applied to a 200 run reading rather than to a failure -- which is why it had to be added at all: a run reading is not an error envelope and carried no detail object. It classifies; it never identifies. No path, object key, URL, credential, token, prompt, payload or host appears here, and no error code was added to the catalog for it.",
            "maxProperties": 16,
            "propertyNames": {
              "pattern": "^[a-z][a-z0-9_]{0,63}$"
            },
            "additionalProperties": {
              "type": [
                "string",
                "number",
                "integer",
                "boolean",
                "null"
              ],
              "maxLength": 256
            }
          },
          "interrupted_reason": {
            "description": "OD-10: a run left `running` by a crash is reconciled to `failed` with an explicit interrupted reason. There is no `interrupted` state.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_finding_count": {
            "type": "integer",
            "minimum": 0
          },
          "diagnostic_observation_count": {
            "type": "integer",
            "minimum": 0,
            "description": "Ungrounded model items rejected by the grounding gate. They are not findings and appear in no finding list."
          },
          "cost_micros": {
            "type": "integer",
            "minimum": 0,
            "description": "What this run spent at the provider, in millionths of the provider currency unit, as an integer. Never a floating-point money value. It is the sum of `cost_micros` over **every** `model_call` row of this run, including every retry attempt, because the per-run budget ceiling is enforced over the same span. Absent when the run made no provider call at all, which is a different fact from a cost of zero."
          },
          "cost_basis": {
            "$ref": "#/components/schemas/CostBasis"
          },
          "model_call_count": {
            "type": "integer",
            "minimum": 0,
            "description": "How many provider calls `cost_micros` sums. Published rather than left to be inferred: a reader who cannot see it cannot tell a run answered first time from one that was retried, and the two spent differently. Present exactly when `cost_micros` is."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "terminal_at": {
            "oneOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          },
          "reaudit_of_run_id": {
            "$ref": "#/components/schemas/RunId"
          }
        }
      },
...
      "ExecutionDispatchStatus": {
        "properties": {
          "paused": {
            "type": "boolean",
            "title": "Paused"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paused",
          "changed_at"
        ],
        "title": "ExecutionDispatchStatus"
      },
...
      "ExecutionJournalPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExecutionJournalEntry"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "title": "ExecutionJournalPage"
      },
      "ExecutionQueueItem": {
        "properties": {
          "job_id": {
            "$ref": "#/components/schemas/JobId"
          },
          "run_id": {
            "$ref": "#/components/schemas/RunId"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "available_at": {
            "type": "string",
            "format": "date-time",
            "title": "Available At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "run_id",
          "state",
          "priority",
          "created_at",
          "available_at"
        ],
        "title": "ExecutionQueueItem"
      },
      "ExecutionQueuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExecutionQueueItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "paused"
        ],
        "title": "ExecutionQueuePage"
      },
...
      "SetExecutionPausedRequest": {
        "properties": {
          "paused": {
            "type": "boolean",
            "title": "Paused"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paused"
        ],
        "title": "SetExecutionPausedRequest"
      },
      "SetJobPriorityRequest": {
        "properties": {
          "priority": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Priority"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "priority"
        ],
        "title": "SetJobPriorityRequest"
      }
    }
  }
}